cfpa2_demo/requirements.txt
numpy>=1.24
matplotlib>=3.7
PyYAML>=6.0
pandas>=2.0
imageio>=2.31
